Adding config parameter: eunit_exclude_patterns
In order to make it possible to filter out files that shall not
be treated by Eunit. The following parameter specifies a list
of regexps that are tried on each file that rebar finds. If a
regexp match, that particular file will be excluded from the
set of files that Eunit will operate on. Example:

  {eunit_exlude_patterns, ["test/.*"]}.

The above will exclude any files in the test directory.

Also, removed the hardcoded {d,TEST} setting for eunit since
it conflicted with local definitions. If needed, it should
be set via the eunit_compile_opts parameter.